add confirm message to clear schedule function

+++ frontend/src/pages/CibusPlanningPage/CibusPlanningPage.js
@@ -88,21 +88,28 @@ const CibusPlanning = (props) => {
   };
 
   const clearSchedule = async (schedule) => {
-    try {
-      let response = await axios.delete(
-        `http://127.0.0.1:8000/api/schedules/${schedule.id}/`,
-        {
-          headers: {
-            Authorization: "Bearer " + token,
-          },
-        }
-      );
-      // getScheduledMeals(schedule);
-      setScheduledMeals(null)
-      console.log(response);
-    } catch (error) {
-      console.log(error.message);
+    if (
+      !window.confirm(
+        "Are you sure you want to clear this schedule? This action cannot be undone."
+      )
+    ) {
+      return null;
     }
+    try {
+        let response = await axios.delete(
+          `http://127.0.0.1:8000/api/schedules/${schedule.id}/`,
+          {
+            headers: {
+              Authorization: "Bearer " + token,
+            },
+          }
+        );
+        // getScheduledMeals(schedule);
+        setScheduledMeals(null)
+        console.log(response);
+      } catch (error) {
+        console.log(error.message);
+      }
   };
 
   return (
